inline evm factory
